# Add 'cpu-count' switch to sys cpu command for retrieving CPU count