# Description

closes #14470

This PR adds a switch to the `sys cpu` command named `--cpu-count/-c` which only retrieves the amount of cpus you have in your system. It's significantly faster than just running `sys cpu | length` because `sys cpu` gets all the sysinfo information about your cpu.

Getting the cpu count is sometimes helpful with `par-each` and other things.

### Before
```nushell
❯ sys cpu
╭─#─┬─name─┬────brand─────┬─freq─┬─cpu_usage─┬───load_average───┬─vendor_id─╮
│ 0 │ 1    │ Apple M1 Max │ 3228 │      9.80 │ 1.46, 1.47, 1.88 │ Apple     │
│ 1 │ 2    │ Apple M1 Max │ 3228 │     10.00 │ 1.46, 1.47, 1.88 │ Apple     │
│ 2 │ 3    │ Apple M1 Max │ 3228 │      5.00 │ 1.46, 1.47, 1.88 │ Apple     │
│ 3 │ 4    │ Apple M1 Max │ 3228 │      2.40 │ 1.46, 1.47, 1.88 │ Apple     │
│ 4 │ 5    │ Apple M1 Max │ 3228 │      2.40 │ 1.46, 1.47, 1.88 │ Apple     │
│ 5 │ 6    │ Apple M1 Max │ 3228 │      0.00 │ 1.46, 1.47, 1.88 │ Apple     │
│ 6 │ 7    │ Apple M1 Max │ 3228 │      4.90 │ 1.46, 1.47, 1.88 │ Apple     │
│ 7 │ 8    │ Apple M1 Max │ 3228 │      0.00 │ 1.46, 1.47, 1.88 │ Apple     │
│ 8 │ 9    │ Apple M1 Max │ 3228 │      4.80 │ 1.46, 1.47, 1.88 │ Apple     │
│ 9 │ 10   │ Apple M1 Max │ 3228 │      2.40 │ 1.46, 1.47, 1.88 │ Apple     │
╰─#─┴─name─┴────brand─────┴─freq─┴─cpu_usage─┴───load_average───┴─vendor_id─╯
❯ use std/bench
❯ bench { sys cpu } --rounds 100
╭───────┬───────────────────╮
│ mean  │ 405ms 401µs 45ns  │
│ min   │ 401ms 474µs 917ns │
│ max   │ 408ms 459µs 459ns │
│ std   │ 1ms 564µs 180ns   │
│ times │ [list 100 items]  │
╰───────┴───────────────────╯
```
and with nproc
```nushell
❯ nproc
10
❯ use std/bench
❯ bench { ^nproc } --rounds 100
╭───────┬──────────────────╮
│ mean  │ 4ms 847µs 6ns    │
│ min   │ 4ms 387µs 125ns  │
│ max   │ 17ms 635µs 666ns │
│ std   │ 1ms 508µs 94ns   │
│ times │ [list 100 items] │
╰───────┴──────────────────╯
```

### After
```nushell
❯ sys cpu -c
10
❯ use std/bench
❯ bench { sys cpu -c } --rounds 100
╭───────┬──────────────────╮
│ mean  │ 248µs 776ns      │
│ min   │ 211µs 792ns      │
│ max   │ 541µs 584ns      │
│ std   │ 42µs 695ns       │
│ times │ [list 100 items] │
╰───────┴──────────────────╯
```

# User-Facing Changes
<!-- List of all changes that impact the user experience here. This helps us keep track of breaking changes. -->

# Tests + Formatting
<!--
Don't forget to add tests that cover your changes.

Make sure you've run and fixed any issues with these commands:

- `cargo fmt --all -- --check` to check standard code formatting (`cargo fmt --all` applies these changes)
- `cargo clippy --workspace -- -D warnings -D clippy::unwrap_used` to check that you're using the standard code style
- `cargo test --workspace` to check that all tests pass (on Windows make sure to [enable developer mode](https://learn.microsoft.com/en-us/windows/apps/get-started/developer-mode-features-and-debugging))
- `cargo run -- -c "use toolkit.nu; toolkit test stdlib"` to run the tests for the standard library

> **Note**
> from `nushell` you can also use the `toolkit` as follows
> ```bash
> use toolkit.nu  # or use an `env_change` hook to activate it automatically
> toolkit check pr
> ```
-->

# After Submitting
<!-- If your PR had any user-facing changes, update [the documentation](https://github.com/nushell/nushell.github.io) after the PR is merged, if necessary. This will help us keep the docs up to date. -->