@fdncred 2024-12-01
Closing in favor of https://github.com/nushell/nushell/pull/14485